bot/requirements.txt
python-telegram-bot==20.0
httpx==0.24.0
asyncpg==0.27.0
